fix(qemu): only build the boot disk if it doesn't exist

+++ linux/run_qemu.sh
@@ -21,33 +21,38 @@ if [ -z "${ANSIBLE_VAULT_PASSWORD+x}" ]; then
 fi
 tmpdir=tmp # ignored by .gitignore
 mkdir -p $tmpdir
-curl -C - -o $tmpdir/$ISO_PATH "$ISO_URL"
-mkdir -p $tmpdir/config
 
-cat > $tmpdir/config/meta-data <<-EOF
+if [ -f $tmpdir/boot-disk.img ]; then
+   echo Boot disk exists, not rebuilding...
+else
+   curl -C - -o $tmpdir/$ISO_PATH "$ISO_URL"
+   mkdir -p $tmpdir/config
+   cat > $tmpdir/config/meta-data <<-EOF
 local-hostname: $HOSTNAME
 EOF
 
-cat cloud-init/user-data \
-| sed "s/GITHUB_REPO/${GITHUB_REPO/\//\\/}/g" \
-| sed "s/GIT_BRANCH/$GIT_BRANCH/g" \
-| sed "s/ANSIBLE_VAULT_PASSWORD/$ANSIBLE_VAULT_PASSWORD/g" \
-> $tmpdir/config/user-data
+   cat cloud-init/user-data \
+       | sed "s/GITHUB_REPO/${GITHUB_REPO/\//\\/}/g" \
+       | sed "s/GIT_BRANCH/$GIT_BRANCH/g" \
+       | sed "s/ANSIBLE_VAULT_PASSWORD/$ANSIBLE_VAULT_PASSWORD/g" \
+       > $tmpdir/config/user-data
 
-echo > $tmpdir/config/vendor-data
+   echo > $tmpdir/config/vendor-data
 
-rm -f $tmpdir/seed.iso
-hdiutil makehybrid -o $tmpdir/seed.iso -hfs -joliet -iso -default-volume-name cidata $tmpdir/config/
-rm -r $tmpdir/config
+   rm -f $tmpdir/seed.iso
+   hdiutil makehybrid -o $tmpdir/seed.iso -hfs -joliet -iso -default-volume-name cidata $tmpdir/config/
+   rm -r $tmpdir/config
 
-cp $tmpdir/$ISO_PATH $tmpdir/boot-disk.img
-qemu-img resize $tmpdir/boot-disk.img +4G
+   cp $tmpdir/$ISO_PATH $tmpdir/boot-disk.img
+   qemu-img resize $tmpdir/boot-disk.img +4G
+fi
 
 qemu-system-x86_64 \
 	 -net 'nic,model=virtio-net-pci' \
 	 -net 'user,hostfwd=tcp::5555-:22' \
 	 -machine $QEMU_MACHINE \
 	 -cpu $QEMU_CPU \
+     -smp 2 \
 	 -m 1024 \
 	 -nographic \
 	 -hda $tmpdir/boot-disk.img \
